+++ scripts/helpers/animation/token/tokenDodgeOnMiss.js
@@ -1,5 +1,6 @@
 import { getVisibleUsers } from "../../anim.js";
-import { getSetting, MODULE_ID } from "../../misc.js";
+import { MS_TO_SEC, MODULE_ID } from "../../const.js";
+import { getSetting } from "../../misc.js";
 import { getTokenImage } from "./shakeOnDamageToken.js";
 
 const COLOR_FILTER = {
@@ -38,6 +39,29 @@ export async function handleDodgeOnMiss(token, target) {
     }
 }
 
+const CONST = {
+    BOUNCE: {
+        ROTATION: { OPPOSITE: 180, OFFSET: 65 },
+        ANCHOR: { BURST: { x: 0.1, y: 0.5 }, SPARK: { x: 0.9, y: 0.5 } },
+        SIZE_MULT: { SPARK: 1.2 * 2 },
+        DELAY_FINISH: { SPARK: -150 },
+    },
+    DODGE: {
+        START_RAY: {
+            x: 0,
+            y: 0,
+        },
+        ANGLE: {
+            ADJUST: {
+                RADIANS: Math.PI / 2,
+            },
+        },
+        DISTANCE: 0.6,
+    },
+};
+
+const direction = () => Sequencer.Helpers.random_array_element([1, -1]);
+
 /**
  * Dodge when attacked
  * @param {*} token Attack token
@@ -45,22 +69,22 @@ export async function handleDodgeOnMiss(token, target) {
  */
 async function dodgeOnMiss(token, target) {
     const distance = getSetting("dodge-on-miss.distance");
-    const duration = getSetting("dodge-on-miss.duration") * 1000;
-    const delay = getSetting("dodge-on-miss.delay") * 1000;
+    const duration = getSetting("dodge-on-miss.duration") * MS_TO_SEC;
+    const delay = getSetting("dodge-on-miss.delay") * MS_TO_SEC;
     const directionRay = new Ray(token.center, target.center);
     const { dx, dy } = Ray.fromAngle(
-        0,
-        0,
-        directionRay.angle + Math.PI / 2,
-        Sequencer.Helpers.random_array_element([1, -1])
+        CONST.DODGE.START_RAY.x,
+        CONST.DODGE.START_RAY.y,
+        directionRay.angle + CONST.DODGE.ANGLE.ADJUST.RADIANS,
+        direction
     );
     const loopBody = { duration: duration / 2, gridUnits: true };
     const scale = (target.document.width + target.document.height) / 2;
-    const rayDistance = Math.max(scale * 0.6, 1) * distance;
+    const rayDistance = Math.max(scale * CONST.DODGE.DISTANCE, 1) * distance;
     const position = { x: dx * rayDistance, y: dy * rayDistance };
     const users = getVisibleUsers(target);
 
-    new Sequence({moduleName: game.modules.get(MODULE_ID).title})
+    new Sequence({ moduleName: game.modules.get(MODULE_ID).title })
         .animation()
         .delay(delay)
         .on(target)
@@ -70,7 +94,7 @@ async function dodgeOnMiss(token, target) {
         .atLocation(target)
         .file(getTokenImage(target?.document))
         .scale({ x: target?.document?.texture?.scaleX ?? 0, y: target?.document?.texture?.scaleY ?? 0 })
-        .scaleToObject(1)
+        .scaleToObject()
         .animateProperty("spriteContainer", "position.x", {
             from: 0,
             to: position.x,
@@ -97,49 +121,51 @@ async function dodgeOnMiss(token, target) {
             fromEnd: true,
             ...loopBody,
         })
-        .waitUntilFinished(-200)
         .duration(duration)
         .forUsers(users)
+        .waitUntilFinished(-200)
         .animation()
         .on(target)
         .waitUntilFinished()
         .opacity(1)
-        .play({preload: true });
+        .play({ preload: true });
 }
 
 async function bounceOffTarget(token, target, filter = {}) {
-    const delay = getSetting("dodge-on-miss.delay") * 1000;
+    const delay = getSetting("dodge-on-miss.delay") * MS_TO_SEC;
     const users = getVisibleUsers(target);
-    const sizeMultiplier = 1.2;
 
-    new Sequence({moduleName: game.modules.get(MODULE_ID).title})
-        .effect() //Burst FX
+    new Sequence({ moduleName: game.modules.get(MODULE_ID).title })
+        //Burst FX
+        .effect()
         .delay(delay)
         .atLocation(target)
-        .rotateTowards(token, { rotationOffset: 180 })
+        .rotateTowards(token, { rotationOffset: CONST.BOUNCE.ROTATION.OPPOSITE })
         .file("jb2a.impact.008.orange")
         .filter("ColorMatrix", filter)
-        .scaleToObject(1)
-        .anchor({ x: 0.1, y: 0.5 })
+        .scaleToObject()
+        .anchor(CONST.ANCHOR.BURST)
         .mirrorX()
-        .waitUntilFinished(-150)
+        .waitUntilFinished(CONST.DELAY_FINISH.SPARK)
         .forUsers(users)
+        //Spark Left
         .effect()
         .atLocation(target)
         .file("jb2a.melee_generic.piercing.two_handed")
         .filter("ColorMatrix", filter)
-        .rotateTowards(token, { rotationOffset: 180 - 65 })
-        .scaleToObject(1 * 2)
+        .rotateTowards(token, { rotationOffset: CONST.BOUNCE.ROTATION.OPPOSITE - CONST.BOUNCE.ROTATION.OFFSET })
+        .scaleToObject(CONST.BOUNCE.SIZE_MULT.SPARK)
         .mirrorY()
-        .spriteAnchor({ x: 0.9, y: 0.5 })
+        .spriteAnchor(CONST.ANCHOR.SPARK)
         .forUsers(users)
+        //Spark Right
         .effect()
         .atLocation(target)
         .file("jb2a.melee_generic.piercing.two_handed")
         .filter("ColorMatrix", filter)
-        .rotateTowards(token, { rotationOffset: 180 + 65 })
-        .scaleToObject(sizeMultiplier * 2)
-        .spriteAnchor({ x: 0.9, y: 0.5 })
+        .rotateTowards(token, { rotationOffset: CONST.BOUNCE.ROTATION.OPPOSITE + CONST.BOUNCE.ROTATION.OFFSET })
+        .scaleToObject(CONST.BOUNCE.SIZE_MULT.SPARK)
+        .spriteAnchor(CONST.ANCHOR.SPARK)
         .forUsers(users)
-        .play({preload: true });
+        .play({ preload: true });
 }