+++ scripts/helpers/const.js
@@ -1,0 +1,17 @@
+//const { ApplicationV2, HandlebarsApplicationMixin } = foundry.applications.api;
+export const MODULE_ID = "pf2e-rpg-numbers";
+
+export const KOFI_MESSAGE = [
+    "Support Me :P",
+    "Buy Me a Coffee",
+    "Support My Work",
+    "Fuel My Creativity",
+    "Coffee = Motivation",
+    "Send a Coffee Boost",
+    "Pay your tribute",
+    "Treat Me to a Brew",
+    "Be the coolest person you know",
+    "Fuel my Paizo addiction",
+];
+
+export const MS_TO_SEC = 1000;