+++ scripts/helpers/misc.js
@@ -1,18 +1,4 @@
-//const { ApplicationV2, HandlebarsApplicationMixin } = foundry.applications.api;
-export const MODULE_ID = "pf2e-rpg-numbers";
-
-export const KOFI_MESSAGE = [
-    "Support Me :P",
-    "Buy Me a Coffee",
-    "Support My Work",
-    "Fuel My Creativity",
-    "Coffee = Motivation",
-    "Send a Coffee Boost",
-    "Pay your tribute",
-    "Treat Me to a Brew",
-    "Be the coolest person you know",
-    "Fuel my Paizo addiction",
-];
+import { MODULE_ID } from "./const.js";
 
 export function debugLog(data, context = "") {
     if (getSetting("debug-mode")) console.log(`PF2E-RPG-#s: ${context}`, data);